+++ 0x02-functions_nested_loops/main.h
@@ -15,6 +15,6 @@ int _islower(char c);
 int _isalpha(int c);
 int print_sign(int x);
 int _abs(int x);
-int print_last_digit(int x);
+int print_last_digit(long x);
 
 #endif
